style(p28.4): two clippy findings, and an arm that tested the compiler rather than the order (P28.4)
`cargo clippy --workspace --all-targets` with `RUSTFLAGS=-D warnings`, run last
per the machine note. Two findings, both in this batch's own test code:

* `READBACK_LATENCY_FRAMES as u64` — the constant is already `u64`;
* `a_shadow_speculation_ranks_below_every_proved_page` compared two `const`s,
  which clippy const-folds and is right to. An arm over two constants cannot
  fail at runtime, so it tested the compiler rather than the rank.

The second is rewritten rather than allowed: what has to be true is that
`inf_stream::normalize` — the one pipeline both slot-pool consumers run — puts
a proved page ahead of a speculated one, offered speculation-first and at the
LOW address so payload order and lane order are opposed; and that one address
wanted by both classes resolves to the PROVED lane, which is the claim that
matters on the frames the predictor guesses right. Both are runtime facts about
the shipped function.

Clippy exit 0, zero warnings. `cargo fmt --all --check` clean. The affected
targets re-run after the fixes: inf-render lib 439, whip_pan 7, stream_arbiter
8, cluster_pages 6.

+++ crates/inf-render/tests/whip_pan.rs
@@ -304,8 +304,8 @@ fn run(predict: Option<u32>, pages: u64) -> Arm {
         // reads at tick `t` was written at `t − READBACK_LATENCY_FRAMES`, and it
         // could only mark surfaces that were visible then. Both halves of that
         // are the lag the predictor exists to close.
-        if tick >= READBACK_LATENCY_FRAMES as u64 {
-            let past = whip_view(tick - READBACK_LATENCY_FRAMES as u64);
+        if tick >= READBACK_LATENCY_FRAMES {
+            let past = whip_view(tick - READBACK_LATENCY_FRAMES);
             wants.extend(
                 justified_tiles(lib.residency(), &past, &cov)
                     .into_iter()
